Refactor of API
